last changes

+++ src/Controller/EmployeeController.php
@@ -3,29 +3,68 @@
 namespace App\Controller;
 
 use App\Entity\Employee;
+use App\Repository\EmployeeRepository;
 use App\Request\EmployeeDto;
 use Doctrine\ORM\EntityManagerInterface;
+use Nelmio\ApiDocBundle\Annotation\Model;
+use OpenApi\Attributes as OA;
 use Symfony\Bridge\Doctrine\Attribute\MapEntity;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Attribute\MapRequestPayload;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Serializer\Encoder\JsonEncoder;
+use Symfony\Component\Serializer\Normalizer\AbstractObjectNormalizer;
+use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 #[Route(path: '/api/employee', name: 'app_employee_')]
 class EmployeeController extends AbstractController
 {
-    public function __construct(private EntityManagerInterface $em)
-    {
+    public function __construct(
+        private EntityManagerInterface $em,
+        private readonly NormalizerInterface $normalizer,
+        private readonly EmployeeRepository $employeeRepository
+    ) {
     }
 
-    #[Route('/{id\d+}', name: 'get', methods: ['GET'])]
+    #[OA\Response(
+        response: 200,
+        description: 'Returns all employees if any are preserved',
+        content: new OA\JsonContent(
+            type: 'array',
+            items: new OA\Items(ref: new Model(type: Employee::class))
+        )
+    )]
+    #[Route('', name: 'list', methods: ['GET'])]
+    public function all(): JsonResponse
+    {
+        return $this->json($this->employeeRepository->findAll());
+    }
+
+    #[OA\Response(
+        response: 200,
+        description: 'Return specific employee by its id',
+        content: new OA\JsonContent(
+            type: 'array',
+            items: new OA\Items(ref: new Model(type: Employee::class))
+        )
+    )]
+    #[OA\Response(
+        response: 404,
+        description: 'Entity not found',
+    )]
+    #[Route('/{id<\d+>}', name: 'get', methods: ['GET'])]
     public function get(#[MapEntity] Employee $employee): JsonResponse
     {
-        return $this->json($employee, Response::HTTP_OK);
+        return $this->json($this->normalizer->normalize($employee, JsonEncoder::FORMAT, [AbstractObjectNormalizer::PRESERVE_EMPTY_OBJECTS => true]), Response::HTTP_OK);
     }
 
-    #[Route('/', name: 'create', methods: ['POST'])]
+    #[OA\Response(
+        response: 201,
+        description: 'Success',
+    )]
+    #[Route('', name: 'create', methods: ['POST'])]
     public function create(#[MapRequestPayload] EmployeeDto $request): JsonResponse
     {
         $employee = new Employee(
@@ -39,11 +78,19 @@ class EmployeeController extends AbstractController
         $this->em->persist($employee);
         $this->em->flush();
 
-        return $this->json($employee, Response::HTTP_CREATED);
+        return $this->json($this->normalizer->normalize($employee, 'json'), Response::HTTP_CREATED);
     }
 
-    #[Route('/{id\d+}', name: 'update', methods: ['PUT'])]
-    public function update(#[MapRequestPayload] EmployeeDto $request, #[MapEntity] Employee $employee): JsonResponse
+    #[Route('/{id<\d+>}', name: 'update', methods: ['PUT'])]
+    #[OA\Response(
+        response: 404,
+        description: 'Entity not found',
+    )]
+    #[OA\Response(
+        response: 204,
+        description: 'Success',
+    )]
+    public function update(#[MapRequestPayload] EmployeeDto $request, #[MapEntity] Employee $employee, EntityManagerInterface $em): JsonResponse
     {
         $employee->touch()
             ->setFirstName($request->firstName)
@@ -53,16 +100,26 @@ class EmployeeController extends AbstractController
             ->setDateOfEmployment($request->dateOfEmployment);
 
         // thanks to UoW only updated fields of managed entity will be in the final update SQL query.
-        $this->em->flush();
+        $em->flush();
 
         return $this->json(null, Response::HTTP_NO_CONTENT);
     }
 
-    #[Route('/{id\d+}', name: 'delete', methods: ['DELETE'])]
-    public function delete(#[MapEntity] Employee $employee): JsonResponse
+    #[Route('/{id<\d+>}', name: 'delete', methods: ['DELETE'])]
+    #[OA\Response(
+        response: 204,
+        description: 'Success',
+    )]
+    public function delete(int $id): JsonResponse
     {
-        $this->em->remove($employee);
-        $this->em->flush();
+        // we don't want to use here #[MapEntity] Employee
+        // since it give to us non-idempotent DELETE which brakes RFC2616
+        // EntityValueResolver works kinda bad with DELETE
+        $employee = $this->employeeRepository->find($id);
+        if (null !== $employee) {
+            $this->em->remove($employee);
+            $this->em->flush();
+        }
 
         return $this->json(null, Response::HTTP_NO_CONTENT);
     }
